+++ xknx/cemi/__init__.py
@@ -11,4 +11,5 @@ from .cemi_frame import (
     CEMIMPropWriteResponse,
 )
 from .cemi_handler import CEMIHandler
-from .const import CEMIErrorCode, CEMIFlags, CEMIMessageCode
+from .const import CEMIErrorCode, CEMIMessageCode
+from .flags import CEMIFlags, CEMIPriority